add a network connection data provider. create a new folder for metadata code.
